+++ CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] / YYYY-MM-DD
-- First Release
+## [1.0.0] / 2022-05-14
+### Features
+- [ ] Download Bundle From Github
+- [ ] Improve Library Files
+### Added
+- Base Package Files
+- Add Example
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0